Reorganized sources
The different functions of the client and the server have been moved to
separate respective packages called "data", "display" and "main".
Basically, "data" is meant to hold classes that design a custom data
structure, for example LineOutputStream; "display" groups the classes
managing, well, the display part for each; and "main" is the starting
point respectively for the client and for the server.

Signed-off-by: Paul Mabileau <paulmabileau@hotmail.fr>