Fixed resize and added a character count limit
The client window has now an easily resizable window with full control
over the dimensions and the split size, but the bottom input text area
will not be modified when resizing the main window. Also, there is now a
way to set a limit on how many character the client can type in, but not
graphically yet.

Signed-off-by: Paul Mabileau <paulmabileau@hotmail.fr>