Implemented dirty serialization
Moved all the textual communication to serialized String objects over
the network, thus breaking the ability to directly send bytes of
characters to the server using netcat for example. But this
"functionality" will be made available again in the near future.
For now, serialization offers a structure for future development,
because server settings will entirely decide what and how to receive
from the client, without having to change major components of the
classes like what just happened.
This commit introduces new bugs linked to serialization - but will be
corrected soon - and a lack of documentation and uniformisation in the
naming of various methods and attributes.

Signed-off-by: Paul Mabileau <paulmabileau@hotmail.fr>