Added serialization choice for client
Added two radio buttons to the client configuration dialog that allow
choosing between the full clear text and the serialized communication.
The MutableBoolean class has thus been moved to the general data package
in order to access it logically from the new client.data.Settings class.
NB: the result of the choice is not implemented yet, only serialized
communication is activated for now.

Signed-off-by: Paul Mabileau <paulmabileau@hotmail.fr>