Updated binaries to new 'jar' folder
The binaries have been updated to make the clear text / serialization
option available as a preview. The binaries are now in a separate folder
called 'jar' in order to avoid the Eclipse Java IDE to erase the
pre-compiled jars because they are in the 'bin' classpath folder, which
happened previously.

Signed-off-by: Paul Mabileau <paulmabileau@hotmail.fr>