Implemented text length limit rule
Added a new rule to the server's configuration dialog that allows one to
put a limit on how long messages received and retransmitted by the
server can be. This commit only makes it available for the textual
non-serialized communication, as the doing it for the object
communication requires further changes.

Signed-off-by: Paul Mabileau <paulmabileau@hotmail.fr>